general code and quality of life improvements

+++ app/lib/actions.js
@@ -30,12 +30,32 @@ export async function saveOrUpdateDocument(id, title, content, userId) {
   const sanitizedTitle = DOMPurify.sanitize(title);
   const sanitizedContent = DOMPurify.sanitize(content);
 
+  // user needs to be logged in to both update or create
+  if (!userId) {
+    throw new Error("Denied - user is not logged in");
+  }
+
   if (!title || title.trim() === "" || !content || content.trim() === "") {
     throw new Error("Title and/or content cannot be empty");
   }
 
   if (id) {
-    // If an id is provided, update the existing document
+    // If an id is provided, find the existing document
+    const existingDocument = await prisma.document.findUnique({
+      where: {
+        id: id,
+      },
+    });
+
+    if (!existingDocument) {
+      throw new Error("Document not found");
+    }
+
+    if (existingDocument.userId !== userId) {
+      throw new Error("User does not have permission to update this document");
+    }
+
+    // If the user is the owner, update the document
     document = await prisma.document.update({
       where: {
         id: id,
@@ -43,10 +63,10 @@ export async function saveOrUpdateDocument(id, title, content, userId) {
       data: {
         title: sanitizedTitle,
         content: sanitizedContent,
-        userId: userId,
       },
     });
   } else {
+    // Create a new document
     document = await prisma.document.create({
       data: {
         title: sanitizedTitle,
@@ -60,7 +80,19 @@ export async function saveOrUpdateDocument(id, title, content, userId) {
 }
 
 // Delete a document from the database
-export async function deleteDocument(id) {
+export async function deleteDocument(id, userId) {
+  const document = await prisma.document.findUnique({
+    where: {
+      id: id,
+    },
+  });
+  if (!document) {
+    throw new Error("Document not found");
+  }
+  if (document.userId !== userId) {
+    throw new Error("User does not have permission to delete this document");
+  }
+
   await prisma.document.delete({
     where: {
       id: id,
